refactor: Split charts tests into separate modules
